+++ docs/components/content/index.ts
@@ -1,1 +1,2 @@
 export { IframeSwitcher, IframeSwitcherGroup } from "./iframe-switcher";
+export { LazyIframe } from "./lazy-iframe";