+++ packages/v1/react-textarea/CHANGELOG.md
@@ -1,5 +1,23 @@
 # ui
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- Updated dependencies [bf1fc6f]
+  - @copilotkit/shared@1.52.2-next.1
+  - @copilotkit/runtime-client-gql@1.52.2-next.1
+  - @copilotkit/react-core@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [f1eaa4c]
+  - @copilotkit/react-core@1.52.2-next.0
+  - @copilotkit/runtime-client-gql@1.52.2-next.0
+  - @copilotkit/shared@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes