+++ packages/v1/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @copilotkit/runtime-client-gql
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- Updated dependencies [d1f7aba]
+- Updated dependencies [bf1fc6f]
+  - @copilotkit/runtime@1.52.2-next.1
+  - @copilotkit/shared@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime@1.52.2-next.0
+- @copilotkit/shared@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes