+++ packages/v1/shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @copilotkit/shared
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- bf1fc6f: fix: `convertJsonSchemaToZodSchema` drops enum constraints
+
+## 1.52.2-next.0
+
 ## 1.52.1
 
 ## 1.52.1-next.1