+++ packages/v2/voice/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/voice
 
+## 1.52.2-next.1
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.2-next.1
+
+## 1.52.2-next.0
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.2-next.0
+
 ## 1.52.1
 
 ### Patch Changes